Allow configuring relativenumber in EasyTree
This commit adds the ability for the user to specify whether relative
line numbering should be turned on in EasyTree windows. By default it is
disabled (with a feature guard as it is a relatively new Vim feature),
mirroring EasyTree's handling of standard line numbering. It can be
enabled by setting

    let g:easytree_show_relative_line_numbers = 1